Moved CLI methods to dedicated class and improves package layout
