--- src/Application.java
@@ -1,16 +1,0 @@
-/**
- * The type Application.
- */
-public class Application {
-
-    /**
-     * The entry point of application.
-     *
-     * @param args the input arguments
-     */
-    public static void main(String[] args) {
-        ForestFire fireSimulation = new ForestFire();
-        fireSimulation.runSimulation();
-    }
-
-}
